Few small updates:
- Use files_all consistently across datasets.
- Update extract_features.sh to reflect this.
- Separate deepspectrum script for ease of use.
- Tweak to write_filelist()
- Version 2021.09.1

+++ papers/interspeech2021/embeddings.sh
@@ -1,49 +1,39 @@
-#!/bin/sh
+#!/usr/bin/env bash
 
 # This script extracts neural embeddings for all corpora
 
-
-# Change this for DeepSpectrum as required depending on GPU memory
-ds_batch_size=16
+export TF_CPP_MIN_LOG_LEVEL=1
 
 # Change this for VGGish/YAMNet as required depending on GPU memory
 as_batch_size=256
 
 # Change to path to wav2vec models dir
-wav2vec_models_dir=$HOME/src/fairseq/examples/wav2vec/
-
-nmels=128
+wav2vec_models_dir=$HOME/src/fairseq/examples/wav2vec
 
 for corpus in CaFE CREMA-D DEMoS EMO-DB EmoFilm eNTERFACE IEMOCAP JL MSP-IMPROV Portuguese RAVDESS SAVEE ShEMO SmartKom TESS URDU VENEC; do
     # auDeep dataset conversion
-    python scripts/utils/convert_audeep_dataset.py \
-        output/$corpus/audeep.nc \
-        $corpus
+    if [ -f "output/$corpus/audeep.nc" ]; then
+        python scripts/utils/convert_audeep_dataset.py \
+            output/$corpus/audeep.nc \
+            $corpus
+    else
+        echo "auDeep file not found for $corpus"
+    fi
 
-    # DeepSpectrum features
+    # DeepSpectrum dataset conversion
     for net in densenet201 densenet169 densenet121; do
-        echo "Extracting for $corpus using $net"
-
-        deepspectrum features \
-            --extraction-network $net \
-            --feature-layer avg_pool \
-            --no-labels \
-            --mode mel \
-            --frequency-scale mel \
-            --number-of-melbands $nmels \
-            --batch-size $ds_batch_size \
-            datasets/$corpus/wav \
-            --output features/$corpus/$net.csv
-        python scripts/utils/convert.py \
-            features/$corpus/$net.csv \
-            features/$corpus/$net.nc \
-            --corpus $corpus
-        rm features/$corpus/$net.csv
+        if [ -f "features/$corpus/$net.csv" ]; then
+            python scripts/utils/convert.py \
+                features/$corpus/$net.csv \
+                features/$corpus/$net.nc \
+                --corpus $corpus
+            rm features/$corpus/$net.csv
+        fi
     done
 
     # Spectrograms
     python scripts/preprocessing/spectrograms.py \
-        datasets/$corpus/files.txt \
+        datasets/$corpus/files_all.txt \
         --corpus $corpus \
         --labels datasets/$corpus/labels.csv \
         --window_size 0.025 \
@@ -65,17 +55,17 @@ for corpus in CaFE CREMA-D DEMoS EMO-DB EmoFilm eNTERFACE IEMOCAP JL MSP-IMPROV 
 
     # (VQ-)Wav2Vec mean embeddings
     python scripts/preprocessing/wav2vec_features.py \
-        --checkpoint $wav2vec_models_dir/wav2vec_large.pt \
-        datasets/$corpus/files.txt \
+        --checkpoint "$wav2vec_models_dir/wav2vec_large.pt" \
+        datasets/$corpus/files_all.txt \
         features/$corpus/wav2vec.nc
     python scripts/preprocessing/wav2vec_features.py \
-        --checkpoint $wav2vec_models_dir/vq-wav2vec.pt \
-        datasets/$corpus/files.txt \
+        --checkpoint "$wav2vec_models_dir/vq-wav2vec.pt" \
+        datasets/$corpus/files_all.txt \
         features/$corpus/vq-wav2vec.nc
     # Wav2Vec 2.0
     python scripts/preprocessing/wav2vec_features.py \
-        --checkpoint $wav2vec_models_dir/libri960_big.pt \
+        --checkpoint "$wav2vec_models_dir/libri960_big.pt" \
         --type 2 \
-        datasets/$corpus/files.txt \
+        datasets/$corpus/files_all.txt \
         features/$corpus/wav2vec2.nc
 done
